# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "APIs for the Github Issue Notifier Project. https://github.com/issue-notifier",
        "title": "Github Issue-Notifier API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hemakshi Sachdev",
            "email": "sachdev.hemakshi@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
        "/api/v1/login/github/oauth2": {
            "get": {
                "description": "Performs a Github OAuth by using the `code` provided in query param and then uses the received accessToken to fetch user information from Github. If it is a new user then user is also saved in the database. It also creates a user session for future authenticated calls.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Login via Github OAuth",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.UserInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/repositories": {
            "get": {
                "description": "Get complete repository information for all repositories from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get all repositories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Repository"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/repository/{repoID}/update/lastEventAt": {
            "put": {
                "description": "Updates `lastEventAt` time for the given `repoID` in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Updates `lastEventAt` time for the given `repoID`",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID for which `lastEventAt` needs to be updated",
                        "name": "repoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "`lastEventAt` time (with timezone) at which last event for the repository occurred. Format `lastEventAt`: `2006-01-02 15:04:05-07:00`",
                        "name": "lastEventAt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.lastEventAt"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/authenticated": {
            "get": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Get user information from Github for authenticated user via session token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user information from Github for authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.UserInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "security": [
                    {
                        "Github OAuth": []
                    }
                ],
                "description": "Performs the logout function by deleting user session",
                "tags": [
                    "user"
                ],
                "summary": "Logout from user session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Repository": {
            "type": "object",
            "properties": {
                "lastEventAt": {
                    "type": "string"
                },
                "repoID": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                }
            }
        },
...
        "routes.UserInfo": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "avatarImg": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "routes.lastEventAt": {
            "type": "object",
            "properties": {
                "lastEventAt": {
                    "type": "string"
                }
            }
        }
    }
}